requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
arxiv==1.4.8
async-timeout==4.0.3
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.2.2
faiss-cpu==1.7.4
fastapi==0.115.12
feedparser==6.0.11
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.0
greenlet==3.1.1
groq==0.22.0
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.8
httpx==0.27.0
huggingface-hub==0.29.3
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.0.335
langsmith==0.0.92
lxml==5.3.1
MarkupSafe==3.0.2
marshmallow==3.26.1
motor==3.3.2
mpmath==1.3.0
multidict==6.2.0
mypy-extensions==1.0.0
networkx==3.4.2
nltk==3.8.1
numpy==1.26.4
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.1.0
pip==22.3.1
propcache==0.3.1
pyasn1==0.6.1
pydantic==2.11.1
pydantic_core==2.33.0
pymongo==4.6.1
PyMuPDF==1.25.5
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.0.1
sentencepiece==0.2.0
setuptools==65.5.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
starlette==0.46.1
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.50.3
typing_extensions==4.13.0
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
